Re-organize file layout
