+++ packages/ts-transformers/src/internal/decorators/event-options.ts
