+++ packages/ts-transformers/src/internal/decorators/query-all.ts
