+++ packages/ts-transformers/src/internal/decorators/query-assigned-nodes.ts
