+++ packages/ts-transformers/src/internal/decorators/query.ts
