+++ packages/ts-transformers/src/internal/lit-class-context.ts
