+++ packages/ts-transformers/src/internal/lit-file-context.ts
